New sc2reader integration. Refactored processing.
WARNING: Untested and definitely unstable.

Rewrote all the processing into a set of composable plugins. These
will take the place of most of the armyjs_map processing as well as
the macro.py processor. Accordingly armyjs_map has be reimplemented
to take advantage of the post processing from the plugins.

Hopefully they are easier to grok and make improvements to.